[refactor] readCompare 변경에 따른 테스트 수정
